requirements.txt
streamlit>=1.32.0
groq>=0.9.0
openai>=1.30.0
sentence-transformers>=2.7.0
faiss-cpu>=1.12.0
Pillow>=10.0.0
pytesseract>=0.3.10
numpy>=1.26.0
pandas>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sympy>=1.12
matplotlib>=3.7.0
scipy>=1.11.0
tqdm>=4.65.0